Move orientation indicator item to params
